move parser files to parser/ dir
