+++ src/core/Formatter.ts
@@ -3,11 +3,11 @@ import Params from './Params';
 import Tokenizer from './Tokenizer';
 import formatCommaPositions from './formatCommaPositions';
 import formatAliasPositions from './formatAliasPositions';
-import Parser from './Parser';
+import Parser from '../parser/Parser';
 import ExpressionFormatter from './ExpressionFormatter';
 import { indentString } from './config';
 import AliasAs from './AliasAs';
-import { Statement } from './ast';
+import { Statement } from '../parser/ast';
 import Layout, { WS } from './Layout';
 import Indentation from './Indentation';
