+++ src/parser/Parser.ts
@@ -1,4 +1,5 @@
 /* eslint-disable no-cond-assign */
+import { EOF_TOKEN, type Token, TokenType, isToken } from 'src/core/token';
 import {
   AllColumnsAsterisk,
   ArraySubscript,
@@ -13,7 +14,6 @@ import {
   Statement,
   TokenNode,
 } from './ast';
-import { EOF_TOKEN, type Token, TokenType, isToken } from './token';
 
 /**
  * A simple parser that creates a very rudimentary syntax tree.
